Some more no-CGAL (bug candidate 1)
