+++ src/geometry/cgal/Polygon2d-CGAL.cc
@@ -85,26 +85,6 @@ mark_domains(CDT& cdt)
 
 } // namespace Polygon2DCGAL
 
-double Polygon2d::area() const
-{
-  const PolySet *p = tessellate();
-  if (p == nullptr) {
-    return 0;
-  }
-
-  double area = 0.0;
-  for (const auto& poly : p->polygons) {
-    const auto& v1 = poly[0];
-    const auto& v2 = poly[1];
-    const auto& v3 = poly[2];
-    area += 0.5 * (
-      v1.x() * (v2.y() - v3.y())
-      + v2.x() * (v3.y() - v1.y())
-      + v3.x() * (v1.y() - v2.y()));
-  }
-  return area;
-}
-
 /*!
    Triangulates this polygon2d and returns a 2D-in-3D PolySet.
  */
